refactor: 개발자 토큰 API 위치를 auth 아래로 이동
