+++ src/main/java/com/techfork/domain/admin/controller/AdminController.java
@@ -1,14 +1,8 @@
 package com.techfork.domain.admin.controller;
 
-import com.techfork.auth.application.command.AuthCommandService;
-import com.techfork.auth.application.command.input.GenerateDeveloperTokenCommand;
-import com.techfork.auth.application.command.result.DeveloperTokenResult;
-import com.techfork.auth.presentation.DeveloperTokenConverter;
-import com.techfork.auth.presentation.response.DeveloperTokenResponse;
 import com.techfork.domain.source.service.CrawlingService;
 import com.techfork.global.common.code.SuccessCode;
 import com.techfork.global.response.BaseResponse;
-import com.techfork.auth.security.oauth.UserPrincipal;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +13,6 @@ import org.springframework.batch.core.repository.JobExecutionAlreadyRunningExcep
 import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
 import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -32,25 +25,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class AdminController {
 
-    private final AuthCommandService authCommandService;
-    private final DeveloperTokenConverter developerTokenConverter;
     private final JobLauncher jobLauncher;
     private final Job summaryAndEmbeddingJob;
     private final CrawlingService crawlingService;
 
-    @Operation(
-            summary = "개발자 토큰 발급 (ADMIN 전용)",
-            description = "프론트엔드 개발/테스트용 장수명 액세스 토큰(30일)을 발급합니다. ADMIN 권한이 있는 사용자만 접근할 수 있습니다."
-    )
-    @PostMapping("/developer-token")
-    public ResponseEntity<BaseResponse<DeveloperTokenResponse>> generateDeveloperToken(
-            @AuthenticationPrincipal UserPrincipal userPrincipal
-    ) {
-        GenerateDeveloperTokenCommand command = developerTokenConverter.toGenerateDeveloperTokenCommand(userPrincipal.getId());
-        DeveloperTokenResult result = authCommandService.generateDeveloperToken(command);
-        DeveloperTokenResponse response = developerTokenConverter.toDeveloperTokenResponse(result);
-        return BaseResponse.of(SuccessCode.OK, response);
-    }
 
     @Operation(
             summary = "요약 추출 + 임베딩 생성 배치 실행 (ADMIN 전용)",
